Specify the default max memory for surefire.
By default the max memory used by surefire will depend on the system.
this however has several issues, on a machine with lots of RAM but not a
huge amount spare each surefire JVM may by default try and grab 2GB of
ram, causing (when concurrency > 1 is used) the build to run out of
memory.
756MB was chosed as it is a multple of 256MB and was enough for a large
complex project with lots of JenkinsRule type tests.