[JENKINS-37062] servlet 3.1 dance...
Jenkins has required the servlet API since version 2.0

However the depnendency you get in plugins is 2.5 which causes the buid
classpath to contain old signatures and does not flag when you are not
implementing a new required API.

Whilst the original dance of keeping backwards compatability with old
cores was worthwhile 2.0 (and 2.7) are now sufficently old that I do not
beleive this dance is waranted, and it did not work anyway for the above
reasons.  So bump the baseline core to 2.7 and use 3.1.

Configures the enforce rul to not care about the Upper Bounds on the
legacy artifactId.