Don't set stupid defaults.
I doubt any plugin cares about multicast dns discovery so just disable
it rather than giving it a non standard port (which could lead to
announcments on a public wifi network)

ALso reuseForks is true by default so do not set it and allow users to
use the standard -DreuseForks=false without having to change the plugin
cofnig.

On the same note concurrency is only ever used in surefires forkCount -
yet another non standard property.  So stop that nonsense and use
forkCount (which is initialized to concurrency for backwards
compatability)