Drop support for 1.x by default in plugin-pom
Jenkins 2.0+ have used servlet api 3.x and as such attepmting to use
this API would have been cumbersome as the default would only expose 2.4
causing funky compilation.  So this moves us firmly into the 2.x era and
cleans up the use of the servlet API dependencies.

At the same time 2.60 is pretty well adopted and uses java8 so I chose
to set 2.60.1 as the baseline and use java 8 by default for compilation.