Bump the version number to 3.0
Whilst this still works with the 1.x line there may be slightly more
than normal you have to do
eg. set the jenkins.version, java.level and do a dnace with the servlet
API.  So lets be upfront and saay this is not your normal change by
bumping the major version.