Remove powermock and mockito.
The versions specified do not work together (depenadbot í´«í´«)
The powermock version is tied to a mockito version.
if a project want to use mocking let them choose what to do and do not
manage it for them.  Users of mocking frameworks should know much better
their requirements than the plugin-pom