Do not reuse surefire (failsafe) forks
Reusing forks is a cause of flakyness in tests, especially ones with
JenkinsRules, as Jenkins iteslf has many things static that may get left
in a zombie like state in between tests.

At the same time update the surefire systemProperty configuration to the
non deprecated systemPropertyVariables and move the headless propery
defintion there.